Improvements in CommentService & PostController(Entity Modification Logic Service단으로)

+++ src/main/java/com/sidepj/ithurts/service/CommentServiceImpl.java
@@ -1,11 +1,16 @@
 package com.sidepj.ithurts.service;
 
 import com.sidepj.ithurts.domain.Comment;
+import com.sidepj.ithurts.domain.Member;
+import com.sidepj.ithurts.domain.Post;
 import com.sidepj.ithurts.repository.CommentRepository;
+import com.sidepj.ithurts.repository.MemberRepository;
+import com.sidepj.ithurts.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.nio.file.AccessDeniedException;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,9 +20,17 @@ import java.util.Optional;
 public class CommentServiceImpl implements CommentService{
 
     private final CommentRepository commentRepository;
+    private final PostService postService;
+    private final MemberService memberService;
 
     @Override
-    public Comment saveComment(Comment comment) {
+    public Comment saveComment(Long postId, Long memberId , Comment comment) {
+        Member commentAuthor = memberService.getMemberById(memberId);
+        comment.setAuthor(commentAuthor);
+
+        Post post = postService.getPost(postId);
+        comment.setPost(post);
+
         return commentRepository.save(comment);
     }
 
@@ -32,7 +45,23 @@ public class CommentServiceImpl implements CommentService{
     }
 
     @Override
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, Long memberId) throws Exception {
+        Member member = memberService.getMemberById(memberId);
+        Comment commentToBeDeleted = getCommentById(commentId);
+
+        if(!commentToBeDeleted.getAuthor().equals(member)){
+            throw new AccessDeniedException("작성한 회원만이 댓글을 삭제할 수 있습니다");
+        }
+
         commentRepository.removeCommentById(commentId);
+    }
+
+    @Override
+    public Comment getCommentById(Long commentId) {
+        Optional<Comment> findComment = commentRepository.findById(commentId);
+        if(findComment.isEmpty()){
+            throw new IllegalArgumentException("해당 댓글을 찾을 수 없습니다");
+        }
+        return findComment.get();
     }
 }
